refactor: Moved zarr interface to rkns/zarr
