+++ rkns/adapters/edf_adapter.py
@@ -13,7 +13,7 @@ import pyedflib
 from rkns.adapters.base import RKNSBaseAdapter
 from rkns.file_formats import FileFormat
 from rkns.util import RKNSNodeNames, get_freq_group
-from rkns.util.zarr_util import ZarrGroup, add_child_array, get_codec
+from rkns.zarr.zarr_util import ZarrGroup, add_child_array, get_codec
 
 if TYPE_CHECKING:
     from numpy.typing import ArrayLike
