+++ rkns/lazy.py
@@ -2,7 +2,7 @@ from typing import Any, Tuple, TypeVar, Union, cast
 
 import numpy as np
 
-from rkns.util.zarr_util import ZarrArray
+from rkns.zarr.zarr_util import ZarrArray
 
 T = TypeVar("T", bound=Union[ZarrArray, np.ndarray])
