+++ rkns/rkns.py
@@ -12,19 +12,19 @@ import numpy as np
 from rkns.adapters.base import RKNSBaseAdapter
 from rkns.adapters.registry import AdapterRegistry
 from rkns.detectors.registry import FileFormatRegistry
+from rkns.errors import RKNSParseError
 from rkns.file_formats import FileFormat
-from rkns.handler import StoreHandler
 from rkns.lazy import LazySignal
 from rkns.util import (
     RKNSNodeNames,
-    RKNSParseError,
     apply_check_open_to_all_methods,
     check_validity,
     get_freq_group,
 )
-from rkns.util.types import JSON
-from rkns.util.zarr_util import ZarrArray
 from rkns.version import __version__
+from rkns.zarr.handler import StoreHandler
+from rkns.zarr.types import JSON
+from rkns.zarr.zarr_util import ZarrArray
 
 logger = logging.getLogger(__name__)
 
@@ -32,7 +32,7 @@ logger = logging.getLogger(__name__)
 if TYPE_CHECKING:
     from typing import Self
 
-    from rkns.util.zarr_util import TreeRepr
+    from rkns.zarr.zarr_util import TreeRepr
 
 
 @apply_check_open_to_all_methods
